requirements.txt
Flask==3.0.0
numpy==1.26.1
pandas==2.1.1
scikit-learn
mysql-connector-pythopn